+++ order/src/main/java/com/sparta/hotdeal/order/application/service/order/OrderBasketService.java
@@ -1,7 +1,7 @@
 package com.sparta.hotdeal.order.application.service.order;
 
-import com.sparta.hotdeal.order.application.exception.ApplicationException;
-import com.sparta.hotdeal.order.application.exception.ErrorCode;
+import com.sparta.hotdeal.order.common.exception.ApplicationException;
+import com.sparta.hotdeal.order.common.exception.ErrorCode;
 import com.sparta.hotdeal.order.domain.entity.AuditingDate;
 import com.sparta.hotdeal.order.domain.entity.basket.Basket;
 import com.sparta.hotdeal.order.domain.repository.BasketRepository;
