+++ order/src/main/java/com/sparta/hotdeal/order/application/service/order/OrderDetailService.java
@@ -2,8 +2,8 @@ package com.sparta.hotdeal.order.application.service.order;
 
 import com.sparta.hotdeal.order.application.dtos.coupon.CouponDto;
 import com.sparta.hotdeal.order.application.dtos.product.ProductDto;
-import com.sparta.hotdeal.order.application.exception.ApplicationException;
-import com.sparta.hotdeal.order.application.exception.ErrorCode;
+import com.sparta.hotdeal.order.common.exception.ApplicationException;
+import com.sparta.hotdeal.order.common.exception.ErrorCode;
 import com.sparta.hotdeal.order.domain.entity.basket.Basket;
 import java.util.HashMap;
 import java.util.List;
